requirements.txt
PyOpenGL 
glfw 
imgui 
Pillow 
numpy
opencv-python